==> 01_simulation.Rmd <==
---
title: "01_simulation"
author: "randy"
date: '2022-03-08'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(MASS)
library(here)
library(janitor)
library(readr)
library(nlme)
library(data.table)
```

```{r}
names(full)

full<- read_csv("data/epic/Reg_Encounters.csv") %>% View()
  select(cffidno, VisitAge) %>%
  na.omit() %>%
  group_by(cffidno) %>%
  arrange(VisitAge) %>% 
  slice(1L) 
## we just chop the age 
## 

hist(full$VisitAge)
```


the quantile is between 3 and 8

```{r}
demog <- here::here("data", "epic", "Demog.csv") %>%
  read.csv() %>%
  janitor::clean_names() %>%
  dplyr::select(id = cffidno, sex, 
         ethnic, mutation1, 
         mutation2, race)

epic <- here::here("data", "epic",
                   "registration_age_min_3_4.csv") %>%
  read.csv(row.name = 1) %>%
  janitor::clean_names()

## if there is one unkwon or missing just treated as one allel.

data <- left_join(epic, demog, by = "id") %>%
  mutate(genotype = case_when(mutation1 == "F508del" & mutation2 == "F508del" ~ "del/del",
                              mutation1 != "F508del" & mutation2 == "F508del" ~ "mut/del",
                              mutation2 != "F508del" & mutation1 == "F508del" ~ "mut/del",
                              mutation1 != "F508del" & mutation2 != "F508del" ~ "mut/mut"))
View(data)
names(data)


data %>%
  ggplot(aes(age, ht, 
             group = sex, 
             color = sex)) +
  geom_smooth() +
  theme_bw()

?geom_smooth

table(data$race)
table(data$ethnic)
```


```{r}
library(splines)
mod_pl <- lme(ht ~ bs(age, knots = c(5, 10, 15), degree = 1) +
                sex + race + genotype,
              random = ~ 1 | id,
              data = data)

broom.mixed::tidy(mod_pl)
summary(mod_pl)

broom.mixed::augment(mod_pl) %>% 
  dplyr::select(age, .fixed) %>%
  arrange(age) %>%
  ggplot(aes(age, .fixed)) +
  geom_line()


broom.mixed::tidy(mod_pl)
```

```{r}
mod_cl <- lme(ht ~ bs(age, knots = c(10), 
                      degree = 3) + sex + race + genotype,
              random = ~ 1 | id,
              data = data)

broom.mixed::tidy(mod_cl)
summary(mod_cl)

broom.mixed::augment(mod_cl) %>% 
  dplyr::select(age, .fixed) %>%
  arrange(age) %>%
  ggplot(aes(age, .fixed)) +
  geom_line()
```



```{r}
library(mgcv)

mod_gamm <- gam(ht ~ s(age, bs = "cr") + s(id, bs = "re") + sex + race + genotype,
              data = data)
summary(mod_gamm)

plot(mod_gamm)
```



```{r}
## generate spline --------------------------------------------------
genSpline <- function(x, knots, degree, theta) {

  basis <- bs(x = x, knots = knots, degree = degree,
              intercept = TRUE)

  y.spline <- basis %*% theta

  dt <- data.table(x, y.spline = as.vector(y.spline))

  return(list(dt = dt, basis = basis, knots = knots))

}


## plot.basis -------------------------------------------------------
plot.basis <- function(basisdata) {
  
  dtbasis <- as.data.table(basisdata$basis)
  dtbasis[, x := seq(0, 1, length.out = .N)]
  dtmelt <- melt(data = dtbasis, id = "x", 
                 variable.name = "basis", variable.factor = TRUE)

  ggplot(data=dtmelt, aes(x=x, y=value, group = basis)) +
      geom_line(aes(color=basis), size = 1) +
      theme(legend.position = "none") +
      scale_x_continuous(limits = c(0, 1), 
                         breaks = c(0, basisdata$knots, 1)) +
      theme(panel.grid.minor = element_blank())
}

## plot spline ---------------------------------------------------------
plot.spline <- function(basisdata, points = FALSE) {

    p <- ggplot(data = basisdata$dt)
    
    if (points) p <- p + geom_point(aes(x=x, y = y), color = "grey75")  
      
    p <- p + 
      geom_line(aes(x = x, y = y.spline), color = "red", size = 1) +
      # scale_y_continuous(limits = c(0, 1)) +
      # scale_x_continuous(limits = c(0, 1), breaks = knots) +
      theme(panel.grid.minor = element_blank())
       
   return(p)
  
}
```

```{r}
x <- seq(5, 20, length.out = 1000)
knots <- c(5, 10, 15)
theta <- mod_pl$coefficients$fixed[1:5]
sdata <- genSpline(x, knots, 1, theta)
plot.basis(sdata)
plot(sdata$dt$x, sdata$dt$y.spline)
plot.spline(sdata)
```


```{r}
x <- seq(5, 19, length.out = 1000)
knots <- c(5, 10, 15)
theta <- mod_cl$coefficients$fixed %>%
  t() %>% 
  as.data.frame() %>%
  dplyr::select(contains(c("Intercept", "bs"))) %>%
  t() 
theta

broom.mixed::tidy(mod_cl)
summary(mod_cl)

sdata <- genSpline(x, knots, 3, theta)
plot(sdata$dt$x, sdata$dt$y.spline)
plot.spline(sdata)

sdata$dt

```


```{r}
set.seed(555)
n <- 100
m <- rpois(n, 10)

broom.mixed::tidy(mod_pl)
id <- 1:n %>% map2(m, ~rep(.x, .y)) %>% unlist()
sex <- rbinom(n, 1, 0.5) %>% map2(m, ~rep(.x, .y)) %>% unlist()
race <- sample(1:6, size = n, replace = TRUE) %>% map2(m, ~rep(.x, .y)) %>% unlist()
ran_int <- rnorm(n, mean  = 0, sd = 5.5) %>% map2(m, ~rep(.x, .y)) %>% unlist()
res <- rnorm(m * n, mean = 0, sd = 2.79) 

spline <- sdata$dt 
data_spl <- map_df(m, ~sample_n(spline, .) %>% arrange(x))

data <- cbind(id, data_spl, sex, race, ran_int, res) %>%
  as.data.frame() %>%
  mutate(y_exp = 91.95 + 0.78 * sex +0.23 * race + ran_int,
         y_obs = y_exp + y.spline + res)
```


```{r fig.height=10, fig.width=10}
plot <- data %>%
  ggplot(aes(x, y_obs, group = as.factor(id), color = as.factor(id))) +
  geom_line() +
  theme_bw() +
  theme(legend.position = "none") +
  facet_wrap(id)
plot
```

- interaction terms with age
- pull out the basis functions
- use point estimates
