==> 03_models.Rmd <==
---
title: "02_models"
author: "Randy"
date: "7/12/2021"
output: html_document
---


```{r setup, include = FALSE}
knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
library("tidymodels")
library("tidyverse")
library("GLMMadaptive")
library("nlme")
library("lme4")

library("here")
library("janitor")
library("gtsummary")
library("flextable")
library("knitr")
library("tinytex")
library("bookdown")

library("HRW")
library("lattice")

library("mgcv")
library("refund")
library("fields")

trellis.par.set(par.xlab.text = list(cex = 1.1),
                par.ylab.text = list(cex = 1.1),
                axis.text = list(cex = 0.8), 
                strip.border = list(cex = 0.8))
```


## loading fev_data
```{r message=FALSE, warning=FALSE}
fev_2020 <- here::here("data", "fev_data_2020.csv") %>% 
  read_csv() %>%
  janitor::clean_names() %>%
  filter(id != 342) %>% View()
# View(fev_2020)
# names(fev_2020)
# str(fev_2020)

## dataset for table1
data0_fev <- fev_2020 %>% 
  dplyr::select(Age_first = age_first,
                Age_last = age_last,
                Age_change = agen,
                # Age_change = age_yrs,
                Repeat = n_visits,
                Sex = sexf,
                Height = height,
                Weight = weight,
                FEV = fev,
                Mutation = mut,
                id, 
                FU = fu) %>%
  mutate(Mutation = case_when(Mutation == 1 ~ "Type I",
                              Mutation == 2 ~ "Type II",
                              Mutation == 3 ~ "Type III"),
         ## Tue Jul 06 11:32:17 2021 ------------------------------------------- 
         ## ask about the sex label
         Sex = case_when(Sex == 1 ~ "Male",
                         Sex == 0 ~ "Female"))


View(data0_fev)

fixed <- model.matrix(~ Sex * Age_change, 
                      data = data0_fev)
random <- model.matrix(~ 1, 
                       data = data0_fev)
```


```{r}
trainings <- data0_fev %>%
  group_by(id) %>%
  nest() %>%
  as.data.frame() %>%
  sample_n(170) %>% 
  unnest(cols = c(data))

testings <- data0_fev %>%
  anti_join(training, 
            by = c("id"))

View(trainings)
View(testings)
```


```{r}
# model1 <- mixed_model(fixed = "Weight ~ Sex * Age_change", 
#                       random = "~ 1 | id",
#                       data = trainings,
#                       family = gaussian())


model1 <- nlme::lme(Weight ~ Sex + Age_change,
                    random = ~ 1 | id,
                    data = trainings,
                    na.action=na.exclude)


model2 <- nlme::lme(Weight ~ Sex + Age_change,
                    random = ~ Age_change | id,
                    data = trainings,
                    na.action=na.exclude)
broom.mixed::tidy(model1)
broom.mixed::glance(model1)
broom.mixed::augment(model1)

View(trainings)
is.na.data.frame(trainings)
```


```{r}
test_baseline <- testings %>%
  filter(Age_change == 0)


predict1 <- 
  IndvPred_lme(model1,
    newdata = test_baseline,
    timeVar = "time",
    M = 300,
    times = unique(testings$Age_change),
    return_data = TRUE,
    level = 0.5,
    interval = "prediction",
    seed = 555)

predict2 %>%
  filter(id == 2) %>% View
```




















```{r}
model1 <- mixed_model(fixed = Weight ~ Sex * Age_change, 
                      random = ~ 1 | id,
                      data = trainings,
                      family = poisson())
predict.mixed_model
predict.lm()
predict.lme
```


```{r}
pred1 <- predict(model1, 
                 newdata = testings[testings$Age_change == 0, ],
                 newdata2 = testings[testings$Age_change >= 0, ], 
                 type = "subject_specific",
                 se.fit = TRUE, 
                 return_newdata = TRUE)


sloop::ftype(predict)
# sloop::otype(predict.MixMod)
sloop::s3_get_method(predict.MixMod)
sloop::s3_dispatch(predict())

methods(class = "MixMod")
methods(predict)


UseMethod(predict.MixMod())
```
