==> 04_epic.Rmd <==
---
title: "02_epic"
author: "Randy"
date: "1/4/2022"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("tidyverse")


library("here")
library("janitor")
library("gtsummary")
library("flextable")

library("knitr")
library("tinytex")
library("bookdown")


library("HRW")
library("lattice")
library("mgcv")


library("refund")
library("fields")
```

1.  filter the data with height percentage in 0 to 99.99%
2.  visit number larger than 10

```{r}
epic_hw0 <- here::here("data", "epic", "Reg_Encounters.csv") %>% 
  read_csv(show_col_types = FALSE) %>%
  janitor::clean_names() %>% 
  dplyr::select(id = cffidno,  
                age = visit_age,
                ht, htpct,
                wt, wtpct) %>%
  # group_by(id) %>%
  # summarize(n = n())
  na.omit() %>%
  # Tue Jan 04 11:14:02 2022 ------------------------------
  ## remove partial incorrect observations
  ## keep the individuals
  filter(htpct < 99.99 & htpct > 0,
         ## starting at least from 3
         age >= 3)

View(epic_hw0)
nrow(epic_hw0)

# names(epic)
## currently has 1710 individuals
## used to be 
epic_ind0 <- epic_hw0 %>%
  group_by(id) %>%
  summarize(hmean = mean(ht, na.rm = T),
            hmed = median(ht, na.rm = T),
            wmean = mean(wt, na.rm = T),
            wmed = median(wt, na.rm = T),
            ## this is the time::starting age
            age_min = min(age),
            age_max = max(age),
            age_med = median(age),
            vnum = n()) %>%
  mutate(age_diff = age_max - age_min) %>%
  # Tue Jan 04 09:34:32 2022 ------------------------------
  ## after this step 1761 individuals
  ## with 76497 observations
  ## only select the visit number over 10 times
  filter(vnum >= 10)
  ## after this step is 1664 individuals
  ## 75959 observations

# View(epic_ind)
# nrow(epic_ind)
# sum(epic_ind$vnum)
# nrow(epic_hw)
```

3.  minimal age smaller than 4 3\*. age difference larger than 5???

```{r}
id_age4 <- epic_ind0 %>% 
  filter(age_min <= 4) %>%
  # filter(age_min <= 5) %>%
  ## filter with age4 1370 individuals
  ## filter with age5 1446 individuals
  dplyr::select(id) %>% 
  unlist()

# length(id_age4)

epic_hw1 <- epic_hw0 %>%
  filter(id %in% id_age4)

epic_ind1 <- epic_hw1 %>%
  group_by(id) %>%
  summarize(hmean = mean(ht, na.rm = T),
            hmed = median(ht, na.rm = T),
            wmean = mean(wt, na.rm = T),
            wmed = median(wt, na.rm = T),
            ## this is the time::starting age
            age_min = min(age),
            age_max = max(age),
            age_med = median(age),
            vnum = n()) %>%
  mutate(age_diff = age_max - age_min) %>%
  filter(age_diff >= 5)
  # Tue Jan 04 09:34:32 2022 ------------------------------
  ## only select the visit number over 10 times
  ## 1325 individuals left so far


# Tue Apr 12 11:28:07 2022 ------------------------------
# nrow(epic_ind1)
  # Tue Apr 12 08:23:10 2022 ------------------------------
  ## age difference large than 5
  # age_diff >= 5
  ## 1272 individuals
```

```{r}
epic_ind1 %>% filter(age_diff >= 5) %>% nrow() ## 1272
epic_ind1 %>% filter(age_diff >= 8) %>% nrow() ## 919
epic10_id <- epic_ind1 %>% 
  filter(age_diff >= 10) %>% 
  dplyr::select(id, )
  ## 645
```

```{r "summary"}
# View(epic_10_id)
# summary(epic_ind)

ggplot(data = epic_ind0, 
       aes(vnum)) +
  geom_histogram(aes(y = ..density..),
                 fill = "darkblue", 
                 bins = 50,
                 # alpha = 0.3, 
                 color = "white") +
  geom_density(color = "indianred",
                 size = 1) +
  theme_bw()

ggsave("figure/visiting_times.png")

ggplot(data = epic_ind0, 
       aes(age_min)) +
  geom_histogram(aes(y = ..density..),
                 fill = "darkblue", 
                 bins = 50,
                 # alpha = 0.3, 
                 color = "white") +
  geom_density(color = "indianred",
                 size = 1) +
  theme_bw()
ggsave("figure/beginning_age_times.png")


ggplot(data = epic_ind0, 
       aes(age_diff)) +
  geom_histogram(aes(y = ..density..),
                 fill = "darkblue", 
                 bins = 50,
                 # alpha = 0.3, 
                 color = "white") +
  geom_density(color = "indianred",
                 size = 1) +
  theme_bw()
ggsave("figure/follow_up_time.png")


ggplot(data = epic_ind1, aes(vnum)) +
  geom_histogram(aes(y = ..density..),
                 fill = "darkblue", 
                 bins = 50,
                 # alpha = 0.3, 
                 color = "white") +
  geom_density(color = "indianred",
                 size = 1) +
  theme_bw()
ggsave("figure/visiting_times_study.png")

ggplot(data = epic_ind1,
       aes(age_min)) +
  geom_histogram(aes(y = ..density..),
                 fill = "darkblue", 
                 bins = 50,
                 # alpha = 0.3, 
                 color = "white") +
  geom_density(color = "indianred",
                 size = 1) +
  theme_bw()
ggsave("figure/beginning_age_study.png")


ggplot(data = epic_ind1, 
       aes(age_diff)) +
  geom_histogram(aes(y = ..density..),
                 fill = "darkblue", 
                 bins = 50,
                 # alpha = 0.3, 
                 color = "white") +
  geom_density(color = "indianred",
               size = 1) +
  theme_bw()
ggsave("figure/follow_up_time_study.png")
```



```{r}
# hist(epic_ind1$vnum, breaks = 100)
# hist(epic_ind1$age_min, breaks = 40)
# hist(epic_ind1$age_diff, breaks = 40)
# View(epic_hw)
# View(epic_hw)

nrow(epic_hw1) ## 66188
nrow(epic_ind1) ## 1370
summary(epic_ind1)

filter(epic_ind1)

write.csv(epic_hw1, file = "data/epic/registration_age_min_3_4.csv")
```

<!-- ```{r} -->

<!-- id103687 <- epic_hw %>% -->

<!--   filter(id == 103687)  -->

<!-- id103687 %>% -->

<!--   ggplot(aes(x = age, y = ht)) + -->

<!--   geom_line() + -->

<!--   theme_bw() + -->

<!--   theme(legend.position = "none") -->

<!-- id140463 <- epic_hw %>% -->

<!--   filter(id == 140463)  -->

<!-- id140463 %>% -->

<!--   ggplot(aes(x = age, y = ht)) + -->

<!--   geom_line() + -->

<!--   theme_bw() + -->

<!--   theme(legend.position = "none") -->

<!-- id141036 <- epic_hw %>% -->

<!--   filter(id == 141036)  -->

<!-- id141036 %>% -->

<!--   ggplot(aes(x = age, y = wt)) + -->

<!--   geom_line() + -->

<!--   theme_bw() + -->

<!--   theme(legend.position = "none") -->

<!-- id150448 <- epic_hw %>% -->

<!--   filter(id == 150448)  -->

<!-- id150448 %>% -->

<!--   ggplot(aes(x = age, y = wt)) + -->

<!--   geom_line() + -->

<!--   theme_bw() + -->

<!--   theme(legend.position = "none") -->

<!-- id107803 <- epic_hw %>% -->

<!--   filter(id == 107803)  -->

<!-- id107803 %>% -->

<!--   ggplot(aes(x = age, y = wt)) + -->

<!--   geom_line() + -->

<!--   theme_bw() + -->

<!--   theme(legend.position = "none") -->

<!-- id103614 <- epic_hw %>% -->

<!--   filter(id == 103614)  -->

<!-- id103614 %>% -->

<!--   ggplot(aes(x = age, y = wt)) + -->

<!--   geom_line() + -->

<!--   theme_bw() + -->

<!--   theme(legend.position = "none") -->

<!-- View(id103614) -->

<!-- View(id107803) -->

<!-- View(id140463) -->

<!-- View(id103687) -->

<!-- View(id141036) -->

<!-- ``` -->

```{r eval=FALSE, fig.height=30, fig.width=30, include=FALSE}
plot_h <- epic_hw %>%
  ggplot(aes(x = age, y = ht, 
             col = as.factor(id))) +
  geom_line() +
  theme_bw() +
  theme(legend.position = "none")

plot_h +
  facet_wrap("id", nrow = 40)

ggsave("plot_height_age_3_4.pdf")
```

```{r eval=FALSE, fig.height=30, fig.width=30, include=FALSE}
plot_w <- epic_hw %>%
  filter(id %in% epic_ind$id) %>% 
  ggplot(aes(x = age, y = wt, 
             col = as.factor(id))) +
  geom_line() +
  theme_bw() +
  theme(legend.position = "none")

plot_w + 
  facet_wrap("id", nrow = 40)
ggsave("plot_weight_correction.pdf")
```
