==> 04_model_fit.Rmd <==
---
title: "04_model_fit"
author: "randy"
date: '2022-04-05'
output:
  pdf_document: default
  html_document: default
---

```{r "setup", include=FALSE}
knitr::opts_chunk$set()
rm(list=ls())
graphics.off()
```

```{r include=FALSE}
library(tidyverse)
library(MASS)
library(here)
library(janitor)
library(readr)
library(nlme)
library(data.table)
library(gtsummary)
library(flextable)
library(splines)
```


```{r}
data <- here::here("data", "epic_clean_randy.csv") %>%
  read.csv(row.names = 1)

# View(data)
```

```{r}
plot1 <- data %>%
  ggplot(aes(age, ht, 
             group = sex, 
             color = sex)) +
  geom_smooth() +
  xlab("Age") +
  theme_bw()

plot1
```



```{r}

mod1_1_1 <- lme(ht ~ bs(age, knots = c(15), degree = 1, 
                        intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
tidy1_1_1 <- broom.mixed::tidy(mod1_1_1)
glance1_1_1 <- broom.mixed::glance(mod1_1_1) %>%
  mutate(model = "mod1_1_1")

mod1_1_2 <- lme(ht ~ bs(age, knots = c(10, 15), degree = 1) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance1_1_2 <- broom.mixed::glance(mod1_1_2) %>%
  mutate(model = "mod1_1_2")


mod1_1_3 <- lme(ht ~ bs(age, knots = c(10, 15), degree = 1) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance1_1_3 <- broom.mixed::glance(mod1_1_3) %>%
  mutate(model = "mod1_1_3")

mod1_1_4 <- lme(ht ~ bs(age, knots = c(10, 15), degree = 1) +
                sex +  genotype - 1,
              random = ~ 1 + age| id,
              data = data)
glance1_1_4 <- broom.mixed::glance(mod1_1_4) %>%
  mutate(model = "mod1_1_4")

mod1_1_0 <- lme(ht ~ bs(age, degree = 1) +
                sex +  genotype - 1,
              random = ~ 1| id,
              data = data)
glance1_1_0 <- broom.mixed::glance(mod1_1_0) %>%
  mutate(model = "mod1_1_0")
```


```{r}
mod1_2_1 <- lme(ht ~ bs(age, knots = c(15),
                        degree = 3, intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
tidy1_2_1 <- broom.mixed::tidy(mod1_2_1)
glance1_2_1 <- broom.mixed::glance(mod1_2_1) %>%
  mutate(model = "mod1_2_1")

mod1_2_2 <- lme(ht ~ bs(age, knots = c(10, 15), degree = 3, intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance1_2_2 <- broom.mixed::glance(mod1_2_2) %>%
  mutate(model = "mod1_2_2")


mod1_2_3 <- lme(ht ~ bs(age, knots = c(10, 15), degree = 3, intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance1_2_3 <- broom.mixed::glance(mod1_2_3) %>%
  mutate(model = "mod1_2_3")

mod1_2_4 <- lme(ht ~ bs(age, knots = c(10, 15), degree = 3, intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 + age| id,
              data = data)
glance1_2_4 <- broom.mixed::glance(mod1_2_4) %>%
  mutate(model = "mod1_2_4")

mod1_2_0 <- lme(ht ~ bs(age, degree = 3, intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1| id,
              data = data)
glance1_2_0 <- broom.mixed::glance(mod1_2_0) %>%
  mutate(model = "mod1_2_0")
```

```{r}
mod1_3_1 <- lme(ht ~ ns(age, knots = c(10),
                        intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
tidy1_3_1 <- broom.mixed::tidy(mod1_3_1)
glance1_3_1 <- broom.mixed::glance(mod1_3_1) %>%
  mutate(model = "mod1_3_1")

mod1_3_2 <- lme(ht ~ ns(age, knots = c(10, 15), 
                        intercept = FALSE) +
                sex + genotype - 1,
              random = ~ 1 | id,
              data = data)
glance1_3_2 <- broom.mixed::glance(mod1_3_2) %>%
  mutate(model = "mod1_3_2")


mod1_3_3 <- lme(ht ~ ns(age, knots = c(10, 15), intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance1_3_3 <- broom.mixed::glance(mod1_3_3) %>%
  mutate(model = "mod1_3_3")

mod1_3_4 <- lme(ht ~ ns(age, knots = c(10), intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1 + age| id,
              data = data)
glance1_3_4 <- broom.mixed::glance(mod1_3_4) %>%
  mutate(model = "mod1_3_4")

mod1_3_0 <- lme(ht ~ ns(age, intercept = FALSE) +
                sex +  genotype - 1,
              random = ~ 1| id,
              data = data)
glance1_3_0 <- broom.mixed::glance(mod1_3_0) %>%
  mutate(model = "mod1_3_0")
```


```{r}
mod2_1_1 <- lme(ht ~ ns(age, knots = c(15),
                        intercept = FALSE) * sex +
                  genotype - 1,
              random = ~ 1 | id,
              data = data)
tidy2_1_1 <- broom.mixed::tidy(mod2_1_1)

glance2_1_1 <- broom.mixed::glance(mod2_1_1) %>%
  mutate(model = "mod2_1_1")

mod2_1_2 <- lme(ht ~ ns(age, knots = c(10, 15), 
                        intercept = FALSE) * sex + 
                  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance2_1_2 <- broom.mixed::glance(mod2_1_1) %>%
  mutate(model = "mod2_1_2")


mod2_1_3 <- lme(ht ~ ns(age, knots = c(10, 15), 
                        intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance2_1_3 <- broom.mixed::glance(mod2_1_3) %>%
  mutate(model = "mod2_1_3")

mod2_1_4 <- lme(ht ~ ns(age, knots = c(10, 15), 
                        intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1 + age| id,
              data = data)
glance2_1_4 <- broom.mixed::glance(mod2_1_4) %>%
  mutate(model = "mod2_1_4")

mod2_1_0 <- lme(ht ~ ns(age, intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1| id,
              data = data)
glance2_1_0 <- broom.mixed::glance(mod2_1_0) %>%
  mutate(model = "mod2_1_0")
```


```{r}
mod2_2_1 <- lme(ht ~ bs(age, knots = c(15),
                        degree = 3, intercept = FALSE) * sex +
                  genotype - 1,
              random = ~ 1 | id,
              data = data)
tidy2_2_1 <- broom.mixed::tidy(mod2_2_1)

glance2_2_1 <- broom.mixed::glance(mod2_2_1) %>%
  mutate(model = "mod2_2_1")

mod2_2_2 <- lme(ht ~ bs(age, knots = c(10, 15), 
                        degree = 3, intercept = FALSE) * sex + 
                  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance2_2_2 <- broom.mixed::glance(mod2_2_1) %>%
  mutate(model = "mod2_2_2")


mod2_2_3 <- lme(ht ~ bs(age, knots = c(10, 15), 
                        degree = 3, intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1 | id,
              data = data)
glance2_2_3 <- broom.mixed::glance(mod2_2_3) %>%
  mutate(model = "mod2_2_3")

mod2_2_4 <- lme(ht ~ bs(age, knots = c(10, 15), 
                        degree = 3, intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1 + age| id,
              data = data)

glance2_2_4 <- broom.mixed::glance(mod2_2_4) %>%
  mutate(model = "mod2_2_4")

mod2_2_0 <- lme(ht ~ bs(age, degree = 3, intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1| id,
              data = data)
glance2_2_0 <- broom.mixed::glance(mod2_2_0) %>%
  mutate(model = "mod2_2_0")


# predict(mod2_2_4)
```


```{r}
library(mgcv)
mod3_1_1<- gam(ht ~ s(age, bs = "cr") + 
                 s(id, bs = "re") + genotype + sex,
              data = data)

glance3_1_1 <- broom.mixed::glance(mod3_1_1) %>%
  mutate(model = "mod3_1_1") %>%
  select(sigma = deviance, logLik, AIC, BIC, model)


mod3_1_2<- gam(ht ~ s(age, bs = "cr", k =50) + 
                 s(id, bs = "re") + genotype + sex,
              data = data)
glance3_1_2 <- broom.mixed::glance(mod3_1_2) %>%
  mutate(model = "mod3_1_2") %>%
  select(sigma = deviance, logLik, AIC, BIC, model)

```


```{r}
rbind(glance1_1_0, glance1_1_1, glance1_1_2, glance1_1_3, glance1_1_4,
      glance1_2_0, glance1_2_1, glance1_2_2, glance1_2_3, glance1_2_4,
      glance1_3_0, glance1_3_1, glance1_3_2, glance1_3_3, glance1_3_4,
      glance2_1_0, glance2_1_1, glance2_1_2, glance2_1_3, glance2_1_4,
      glance2_2_0, glance2_2_1, glance2_2_2, glance2_2_3, glance2_2_4,
      glance3_1_1, glance3_1_2) %>%
  as.data.frame() %>%
  select(model, everything()) %>%
  arrange(AIC) %>%
  knitr::kable()
```



```{r}
mod2_2_4 <- lme(ht ~ bs(time, knots = c(15), 
                        degree = 3, intercept = FALSE) * sex +  
                  genotype - 1,
              random = ~ 1| id,
              data = data)

glance2_2_4 <- broom.mixed::glance(mod2_2_4) %>%
  mutate(model = "mod2_2_4")
glance2_2_4


augment2_2_4 <- broom.mixed::augment(mod2_2_4) %>%
  mutate(model = "mod2_2_4") 

fitted <- augment2_2_4 %>%
  select(time, .fitted, sex) %>%
  unique()

```



```{r}
plot1 <- 
  ggplot() +
  geom_smooth(data = data, aes(x = time, y = ht, group = sex, color = sex)) +
  geom_smooth(data = fitted, aes(x = time, y = .fitted, group = sex), linetype = 2) +
  xlab("Age") +
  theme_bw()

plot1
```
AIC with one internal knot
321377

AIC with two internal knots
316511

`r 321377 - 316511` 

```{r}
mod2_2_4 <- lme(ht ~ bs(time, knots = c(10, 15), 
                        degree = 3, 
                        intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1| id,
              data = data)

glance2_2_4 <- broom.mixed::glance(mod2_2_4) %>%
  mutate(model = "mod2_2_4")
glance2_2_4


augment2_2_4 <- broom.mixed::augment(mod2_2_4) %>%
  mutate(model = "mod2_2_4") 

fitted <- augment2_2_4 %>%
  select(time, .fitted, sex) %>%
  unique()

plot1 <- 
  ggplot() +
  geom_smooth(data = data, aes(x = time, y = ht, group = sex, color = sex)) +
  geom_smooth(data = fitted, aes(x = time, y = .fitted, group = sex), linetype = 2) +
  xlab("Age") +
  theme_bw()

plot1
```
