==> 05_brokenstick.Rmd <==
---
title: "05_plm"
author: "randy"
date: '2022-04-05'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# citation <- c("tidyverse", "MASS", 
#               "brokenstick", "gamlss", "lme4") %>% 
#   map(citation)
# citation
```


```{r}
library(tidyverse)
library(magrittr)

library(MASS)
library(here)

library(brokenstick)
library(gamlss)
library(lme4)
```


```{r}
train <- here::here("data", "epic_clean_randy.csv") %>%
  read.csv(row.names = 1) %>% 
  filter(group == "training")

test <- here::here("data", "epic_clean_randy.csv") %>%
  read.csv(row.names = 1) %>%
  filter(group == "testing")

base_train <- train %>% 
  group_by(id) %>% 
  arrange(time) %>% 
  slice(1L) 

base_test <- test %>% 
  group_by(id) %>% 
  arrange(time) %>% 
  slice(1L) 

View(base_train)
View(base_test)

# View(train)
# View(test)
```



```{r}
mod01 <- lme(ht ~ bs(time, knots = c(10, 15), 
                        degree = 3, 
                        intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1| id,
              data = train)

broom.mixed::glance(mod01)
## AIC 210324.6	

mod02 <- lme(ht ~ bs(time, knots = c(10), 
                        degree = 3, 
                        intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1| id,
              data = train)
broom.mixed::glance(mod02)
## AIC 210845.4	


# mod03 <- lme(ht ~ bs(time, knots = c(11, 15), 
#                         degree = 3, 
#                         intercept = FALSE) * sex +  genotype - 1,
#               random = ~ 1| id,
#               data = train)
# glance(mod03)
# tidy(mod03)
# ## AIC 210300.7
# attr(mod03$terms, "predvars")

## spline with df = nk + power
mod04 <- lme(ht ~ bs(time,  
                     df = 2 + 3,
                    degree = 3, 
                    intercept = FALSE) * sex +  genotype - 1,
              random = ~ 1| id,
              data = train)
broom.mixed::glance(mod04)
tidy(mod01) %>% knitr::kable()
tidy(mod04) %>% knitr::kable()
## AIC 209881.1
attr(mod04$terms, "predvars")

```

AIC for knots (10, 15) 211719.6	
AIC for knots 1/3 2/3 211378.7


```{r include=FALSE}
## generate spline --------------------------------------------------
genSpline <- function(x, knots, degree, theta) {

  basis <- bs(x = x, 
              knots = knots, 
              degree = degree,
              intercept = TRUE)

  y.spline <- basis %*% theta

  dt <- data.table(x, y.spline = as.vector(y.spline))

  return(list(dt = dt, basis = basis, knots = knots))

}


## plot.basis -------------------------------------------------------
plot.basis <- function(basisdata) {
  
  dtbasis <- as.data.table(basisdata$basis)
  dtbasis[, x := seq(0, 1, length.out = .N)]
  dtmelt <- melt(data = dtbasis, id = "x", 
                 variable.name = "basis", variable.factor = TRUE)

  ggplot(data=dtmelt, aes(x=x, y=value, group = basis)) +
      geom_line(aes(color=basis), size = 1) +
      theme(legend.position = "none") +
      scale_x_continuous(limits = c(0, 1), 
                         breaks = c(0, basisdata$knots, 1)) +
      theme(panel.grid.minor = element_blank())
}

## plot spline ---------------------------------------------------------
plot.spline <- function(basisdata, points = FALSE) {

    p <- ggplot(data = basisdata$dt)
    
    if (points) p <- p + geom_point(aes(x=x, y = y), color = "grey75")  
      
    p <- p + 
      geom_line(aes(x = x, y = y.spline), color = "red", size = 1) +
      # scale_y_continuous(limits = c(0, 1)) +
      # scale_x_continuous(limits = c(0, 1), breaks = knots) +
      theme(panel.grid.minor = element_blank())
       
   return(p)
  
}
```


```{r}
# ?brokenstick
mod1 <- brokenstick(formula = ht ~ time | id,
                    data = train,
                    subjid = train$id,
                    knots = c(10, 15))

est1_knots <- predict(mod1, 
                     x = "knots", 
                     group = train$id)

est1_all <- predict(mod1, 
                   group = train$id)
# ?predict.brokenstick

# Tue Apr 12 09:45:02 2022 ------------------------------
## what are those numbers
# nrow(data) ## 66188
# nrow(est1_all) ## 44084
# nrow(est1_knots) ## 3652
# View(est1_knots)
# View(est1_all)

est1_test <- predict(mod1, newdata = test)
# View(est1_test)
# nrow(est1_test)
# nrow(test)
```


```{r}
mod2 <- brokenstick(formula = ht ~ time + sex + genotype | id,
                    data = train,
                    subjid = train$id,
                    knots = c(10, 15))

est2_knots <- predict(mod2, 
                     x = "knots", 
                     group = train$id)
est2_all <- predict(mod2, 
                   group = train$id)
est2_test <- predict(mod2, newdata = test)
```


```{r}
mod3 <- brokenstick(formula = ht ~ time * sex + genotype | id,
                    data = train,
                    subjid = train$id,
                    knots = c(10, 15))

est3_knots <- predict(mod3, 
                     x = "knots", 
                     group = train$id)

## for all the points we have... 
est3_all <- predict(mod3, 
                   group = train$id)

est3_test <- predict(mod3, newdata = test)
```

## use time of 12 age with trajectory turning points 
## age 10 and sge 12

```{r}
est_test <- cbind(test$ht, est1_test, 
                  est2_test, est3_test) %>%
  as.data.frame() %>%
  dplyr::select("y" = 1,
                "mod1" = 2,
                "mod2" = 3,
                "mod3" = 4) %>%
  mutate(mse1 = (mod1 - y)^2,
         mse2 = (mod2 - y)^2,
         mse3 = (mod3 - y)^2)

colSums(est_test[5:7])
```


```{r}
est1_0 <- est1_knots %>%
  filter(time == 0) %>% 
  dplyr::select(id, .pred) %>%
  mutate(time = 0)

est1_10 <- est1_knots %>%
  filter(time == 10) %>% 
  dplyr::select(id, .pred) %>%
  mutate(time = 10)

est1_15 <- est1_knots %>%
  filter(time == 15) %>% 
  dplyr::select(id, .pred) %>%
  mutate(time = 15)
```

```{r}
train1 <- train %>% 
  dplyr::select(id, time, 
                sex, genotype, ht) %>% 
  distinct() %>%
  mutate(time = round(time, 0)) %>%
  group_by(id, time) %>%
  slice(1L) %>%
  ungroup() %>%
  filter(time %in% c(0, 10, 15)) 


train2 <- left_join(train1, est1_0, c("id", "time")) %>%
  left_join(est1_10, c("id", "time")) %>%
  left_join(est1_15, c("id", "time")) 
  mutate(yhat = coalesce(pre))

View(train1)  
View(train2)
```


## functional data analysis (just to see)

```{r}
epic_hw10 <- here::here("data", "epic",
                      "registration_age_min_3_4.csv") %>%
  read.csv(row.names = 1) %>%
  filter(id %in% unlist(epic10_id))

epic10_train_id <- sample_n(epic10_id, 320) 

epic10_train <- epic_hw10 %>%
  filter(id %in% unlist(epic10_train_id))

epic10_test <- anti_join(epic_hw10, epic10_train)
```

```{r}
fda_wt1 <- mgcv::gam(wt ~ s(age, bs = "cr", k = 50) + s(factor(id), bs = "cr"), 
                  data = epic10_train)

fdm_wt0 <- mgcv::gamm(wt ~ s(age, bs = "cr", k = 50), 
                   random = list(id = ~1),
                  data = epic10_train)

fdm_wt1 <- mgcv::gamm(wt ~ s(age, bs = "cr", k = 50), 
                   random = list(id = ~1 + age),
                  data = epic10_train)

fda_ht1 <- mgcv::gam(ht ~ s(age, bs = "cr", k = 50) + s(id, bs = "cr"), 
                  data = epic_hw10)

fdm_ht1 <- mgcv::gamm(ht ~ s(age, bs = "cr", k = 50), 
                   random = list(id = ~1),
                  data = epic10_train)
plot(fda_ht1)
plot(fdm_ht1$gam)
plot(fda_wt1)
plot(fdm_wt1$gam)
```


```{r}
summary(fda_wt1)
plot(fda_wt1)
summary(fdm_wt0$lme)
summary(fdm_wt1$lme)
```
