==> 05_table1.Rmd <==
---
title: "05_table1"
author: "randy"
date: '2022-03-22'
output:
  pdf_document: default
  html_document: default
---

```{r "setup", include=FALSE}
knitr::opts_chunk$set()
rm(list=ls())
graphics.off()
```

```{r include=FALSE}
library(tidyverse)

library(readr)

library(here)
library(janitor)

library(nlme)
library(MASS)

library(data.table)
library(gtsummary)
library(flextable)
library(splines)
```


```{r}
`%!in%` <- Negate(`%in%`)

demog <- here::here("data", "epic", "Demog.csv") %>%
  read.csv() %>%
  janitor::clean_names() %>%
  dplyr::select(id = cffidno, sex, 
         ethnic, mutation1, 
         mutation2, race) %>%
  # Tue Apr 10 09:14:48 2022 ------------------------------
  mutate(mutation1 = 
           case_when(mutation1 == "" ~ "Unknown",
                     TRUE ~ as.character(mutation1)),
         mutation2 = 
           case_when(mutation2 == "" ~ "Unknown",
                     TRUE ~ as.character(mutation2))) %>%
  mutate(genotype = 
           case_when(mutation1 == "F508del" & 
                       mutation2 == "F508del" ~ "Two alleles F508del",
                     # mutation1 == "Unknown" & 
                     # mutation2 == "F508del" ~ "Unknown",
                     # mutation2 == "Unknown" & 
                     # mutation1 == "F508del" ~ "Unknown",
                     mutation1 %!in% c("Unknown", "F508del") & 
                       mutation2 == "F508del" ~ "One allele F508del",
                     mutation2 %!in% c("Unknown", "F508del") & 
                       mutation1 == "F508del" ~ "One allele F508del",
                     # mutation1 %!in% c("Unknown", "F508del") & 
                     # mutation2 %!in% c("Unknown", "F508del") ~ "Other",
                     TRUE ~ "Others or Unknown")) %>% 
  mutate(genotype = factor(genotype, 
                           levels = c("Two alleles F508del",
                                      "One allele F508del",
                                      "Others or Unknown")))

# View(demog)
# levels(factor(demog$mutation1))
## Two alleles F508del 
## One allele F508del 
## Other or unknown

epic <- here::here("data", "epic",
                   "registration_age_min_3_4.csv") %>%
  read.csv(row.name = 1) %>%
  janitor::clean_names()
```


```{r}
data <- left_join(epic, demog, by = "id") %>%
  mutate(sex = as.factor(sex))

ID <- unique(data$id)
length(ID) / 3
test <- sample(ID, 457, replace = FALSE)

data0 <- data %>%
  mutate(group = 
           case_when(id %in% test ~ "testing",
                     TRUE ~ "training"))
# View(data0)
```

```{r}
data1 <- data0 %>%
  group_by(id, group) %>%
  summarize(age_mean = mean(age),
            age_min = min(age),
            age_max = max(age),
            age_n = length(age),
            visitn = n(),
            h_mean = mean(ht),
            h_max = max(ht),
            h_min = min(ht),
            # Tue Apr  5 11:09:20 2022 ------------------------------
            ## add BMI at the baseline
            w_mean = mean(wt),
            w_max = max(wt),
            w_min = min(wt),
            sex = sex,
            genotype = genotype,
            ethnic = ethnic,
            race = race) %>% 
  ungroup() %>%
  unique()

# data0 <- data %>%
#   group_by(id, sex, ethnic, genotype) %>%
#   nest()


data2 <- full_join(data1, data) %>% 
  as.data.frame() %>%
  mutate(time = age - age_min,
         age_diff = age_max - age_min,
         BMI = wt / (0.1 * ht)^2) 

# head(data2)
write.csv(data2, file = "data/epic_clean_randy.csv")
```




```{r}
table0 <- data1 %>%
  unique() %>%
  dplyr::select(-id) %>%
  mutate(
    ethnic = case_when(ethnic == 1 ~ "Hispanic",
                     ethnic == 2 ~ "Non-Hispanic"),
    race = case_when(race == 1 ~ "White",
                     race != 1 ~ "Other"),
    sex = case_when(sex == "F" ~ "Female",
                    sex == "M" ~ "Male"),
    age_diff = age_max - age_min) %>% 
  dplyr::select(
         Genotype = genotype,
         Gender = sex,
         Race = race,
         Ethnicity = ethnic,
         "Visit number" = visitn,
         "Age mean" = age_mean, 
         "Age baseline" = age_min, 
         "Age final" = age_max, 
         "Follow up years" = age_diff,
         "Height mean" = h_mean, 
         "Height baseline" = h_min,
         "Weight mean" = w_mean,
         "Weight baseline" = w_min) %>% 
  ## select all the variables for table1
  tbl_summary() %>%
  ## just display all the variables in one column
  modify_header(label = "**Variable**") %>%
  # update the column header
  bold_labels() %>%
  italicize_labels() %>%
  as_flex_table() %>%
  flextable::bold(part = "header") %>% 
  ## auto adjust the column widths 
  flextable::autofit()



table1 <- data1 %>%
  unique() %>%
  dplyr::select(-id) %>%
  mutate(
    ethnic = case_when(ethnic == 1 ~ "Hispanic",
                     ethnic == 2 ~ "Non-Hispanic"),
    race = case_when(race == 1 ~ "White",
                     race != 1 ~ "Other"),
    sex = case_when(sex == "F" ~ "Female",
                    sex == "M" ~ "Male"),
    age_diff = age_max - age_min) %>% 
  dplyr::select(group,
                
         Genotype = genotype,
         Gender = sex,
         Race = race,
         Ethnicity = ethnic,
         "Visit number" = visitn,
         "Age mean" = age_mean, 
         "Age baseline" = age_min, 
         "Age final" = age_max, 
         "Follow up years" = age_diff,
         "Height mean" = h_mean, 
         "Height baseline" = h_min,
         "Weight mean" = w_mean,
         "Weight baseline" = w_min) %>% 
  ## select all the variables for table1
  tbl_summary(by = group) %>%
  ## just display all the variables in one column
  modify_header(label = "**Variable**") %>%
  # update the column header
  bold_labels() %>%
  italicize_labels() %>%
  as_flex_table() %>%
  flextable::bold(part = "header") %>% 
  ## auto adjust the column widths 
  flextable::autofit()


```



```{r}
table0
table1

## save pptx -----------------------------------------------
## flextable can be saved directly to powerpoints
flextable::save_as_pptx(
  table1, 
  path = "figure/01_table1.pptx")

library(webshot)
# webshot::install_phantomjs()

flextable::save_as_image(table1, 
              path = "figure/01_table1.png",
              # zoom = 3, 
              # expand = 10, 
              webshot = "webshot")

flextable::save_as_image(table0, 
              path = "figure/01_table0.png",
              # zoom = 3, 
              # expand = 10, 
              webshot = "webshot")
# ?save_as_image
```


2 knots and bspline with cubic terms


- new table1 with two columns

  - include age_min & age_diff
  - time follow up time
  - training and testing
  - as different label for visit times
  - BMI, ht, wt
  - gender, r/eth, genotype

- split the data as 1/3 for testing and 2/3 for training
  - predicted value and the real obs for each subject in the testing set.
  - at least one observation for that individual

- cross validation GCV; 
as extra methodology for the model fitting

- use the predictive (dynamic predcition) as well as the marginal mean

- use the PML methods.
