==> 09_multiple_time_plm.Rmd <==
---
title: "09_multiple_plm"
author: "randy"
date: "2022-05-31"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r include=FALSE}
library("tidyverse")
library("magrittr")

library("gamlss")
library("brokenstick")

library("rjags")
library("JMbayes")

library("lme4")
library("nlme")
library("splines")

library(gridExtra)
library(grid)
library(ggplot2)
library(lattice)


library(parallel)
detectCores()

not_all_na <- function(x) any(!is.na(x))
not_any_na <- function(x) all(!is.na(x))
`%!in%` <- Negate(`%in%`)
```


```{r}
load("data/epic_randy_train_test.Rdata")
```

```{r}
outcome <- "ht"
time <- "time"
id <- "id"
pred_time <- c(2, 4, 6, 8, 10, 12, 14, 16)
knots <- c(5, 10, 12)
dataset <- train
lmm_var <- c("sex", "genotype", "race")
match_time <- 12
match_num <- 15
spline <- "cs"
df_gamlss <- 3
df_gam_sigma <- 1
num_cyc_gam <- 20

## extract the baseline information
dataset_baseline <- dataset %>%
  group_by(eval(id)) %>%
  arrange(eval(time)) %>%
  slice(1L) %>%
  dplyr::select(baseline = eval(outcome),
                # time = eval(time),
                everything()) %>%
  dplyr::select(-time, -wt, -age) %>%
  ungroup()


## match_time must be in the set of knots
if (match_time %!in% knots) {
  stop("matching time points is not in the set of knots!")
}

## information variables
info <- dataset %>%
  dplyr::select(eval(id),
                eval(lmm_var)) %>%
  unique()

formula <- paste0(outcome, "~", time, "|", id)

## fit a brokenstick model
bks <- brokenstick(formula = as.formula(formula),
                   data = dataset,
                   knots = knots)


## This is not used in the later steps -----------------------------------------
### get brokenstick model predictions
# bks_pred_all <- 
#   predict(bks, group = dataset$id)
# dataset_all_long_ind <-
#   ## merge the predicted with the full dataset
#   full_join(dataset,
#             bks_pred_all) %>%
#   dplyr::rename(ht_bks_pred = `.pred`) %>%
#   dplyr::select(eval(time), 
#                 ht_bks_pred,
#                 eval(id), 
#                 eval(outcome)) %>%
#   ## group the data by each individual
#   split(f = .$id,
#         drop = TRUE) %>%
#   ## find everything in long format
#   ## also save the dataset in each given individual
#   map(~ dplyr::select(., eval(time),
#                       ht_bks_pred,
#                       eval(outcome)) %>%
#         remove_rownames())

## brokenstick model for predictions at each knots
bks_pred_knots <- 
  predict(bks, x = pred_time,
          group = dataset$id)
dataset_knots_wide <-
  bks_pred_knots %>%
  select_if(not_all_na) %>%
  # dplyr::select(target = as.character(match_time),
  #               everything()) %>%
  full_join(dataset_baseline, by = "id") %>%
  mutate(time = paste0("time", time)) %>%
  pivot_wider(names_from = time,
              values_from = .pred)
# View(dataset_knots_wide)
# View(dataset_knots_long)

## fit a linear model with the prediction from brokenstick model
## this is model with 17 time points independently
outcomes <- paste0("time", pred_time, collapse = ", ")
formula_lm <- paste0("cbind(", outcomes, ") ~ ", "sex + baseline")
# formula_lm <- "`.pred` ~ time * sex + baseline"
# str(dataset_knots_wide)
lm_bks <- lm(as.formula(formula_lm),
             data = dataset_knots_wide)
(lm_summary <- summary(lm_bks))
# lm_summary

lb_data <- dataset_knots_wide %>%
  mutate(lm_bks_target = predict(lm_bks,
    newdata = .)) %>%
  dplyr::select(lm_bks_target) %>%
  cbind(dataset_knots_long) %>%
  dplyr::select("id", contains("lm_bks_target")) %>%
  as.matrix() %>%
  as.data.frame() %>%
  pivot_longer(cols = contains("lm_bks"),
               names_to = "time",
               values_to = "lm_bks_target") %>%
  separate(time, into = c("remove", "time"), sep = 18) %>%
  dplyr::select(-remove) %>%
  mutate(lm_bks_target = as.numeric(lm_bks_target))

```

try different distance for the 


```{r message=FALSE, warning=FALSE, include=FALSE}
## a list to store the each individual list
result <- array()
id_sub <- list()
pb <- txtProgressBar(min = 0,      # Minimum value of the progress bar
                     max = length(unique(lb_data$id)), # Maximum value of the progress bar
                     style = 3,    # Progress bar style (also available style = 1 and style = 2)
                     width = 50,   # Progress bar width. Defaults to getOption("width")
                     char = "=")   # Character used to create the bar


pdf("plot/people_like_me_all_matching_trajectories.pdf")
for (i in 1:length(unique(lb_data$id))) {
  sbj <- unique(lb_data$id)[i]
  subject <- lb_data %>%
    filter(id == sbj)

  lb_sub <- lb_data %>% 
    transmute(id = id,
              ## more time points for matching
              ## adding the correlation
              time = time,
              diff = abs(lm_bks_target - subject$lm_bks_target)) %>%
    pivot_wider(names_from = "id",
                values_from = "diff") %>% 
    remove_rownames() %>%
    column_to_rownames("time") %>%
    ## using L2 norm
    apply(2, norm, type = "2") %>% 
    ## using Frobenius norm
    # apply(lb_sub, 2, norm, type = "f") %>%
    as.data.frame() %>%
    dplyr::select(diff = 1) %>%
    rownames_to_column("id") %>%
            arrange(diff) %>%
            ## make it reproducible --------------
            ## considering just here --------------
            slice(1:match_num) %>%
            inner_join(dataset, by = "id")
  
   id_sub[[i]] <- unique(lb_sub$id) %>% unlist()

   ## plot out the matchings with the target -------------------------------------
   plot <- ggplot(lb_sub) +
      geom_line(aes(x = time, y = ht,
                    group = id),
                    color = "grey",
                linetype = "dashed") +
      geom_line(data = filter(lb_sub, id == sbj),
                aes(x = time, y = ht),
                color = "darkblue",
                size = 1) +
      ggtitle(sbj) +
      theme_bw()

   print(plot)

  (formula_gamlss <- paste0(outcome, " ~ ", spline, "(", time, ", df = ", df_gamlss, ")"))
  (formula_gamsigma <- paste0("~ ", spline, "(", time, ", df = ", df_gam_sigma, ")"))
  # fit a gamlss model flexible enough to capture the details
  plm <-
    gamlss(as.formula(formula_gamlss),
           sigma.formula = as.formula(formula_gamsigma),
           # nu.formula = ~cs(time^0.1, df=1),
           # tau.formula = ~cs(time^0.5, df=1),
           data = na.omit(lb_sub),
           n.cyc = num_cyc_gam,
           family = NO)

  ref_txt <-
    centiles.pred(plm,
      linetype = c("centiles"),
      xname = "time",
      xvalues = c(0:17),
      cent = c(5, 10, 25,
               50, 75, 90, 95),
      plot = FALSE,
      legend = TRUE) %>%
    dplyr::select(time = 1,
                  q05 = 2,
                  q10 = 3,
                  q25 = 4,
                  q50 = 5,
                  q75 = 6,
                  q90 = 7,
                  q95 = 8) %>%
    mutate(cfint90 = q95 - q05,
           cfint80 = q90 - q10,
           cfint50 = q75 - q25)

  result[i] <- data.frame(results = nest(ref_txt))
  setTxtProgressBar(pb, i)
  
}
close(pb)
# return(list(quantile = result,
#             pred_knot = dataset_knots_wide,
#             pred_all = dataset_all_long_ind,
#             lm_summary = lm_summary))
dev.off()


# lm_summary
# plm_plot

```

```{r}
save(result,
     id_sub,
     dataset_knots_wide,
     dataset_all_long_ind,
     file = "data/people_like_me_prediction_multiple_n15.Rdata")

load("data/people_like_me_prediction_multiple_n15.Rdata")
# result <- map(result, as.data.frame)
# View(result)
```


```{r}
plm_plot <- function(id,
                     quantile = result,
                     observation = dataset_all_long_ind) {
  # quantile = result
  # observation = dataset_all_long_ind
  # View(result[[1]])

  quantile_ind = as.data.frame(quantile[[id]])
  observation_ind = as.data.frame(observation[[id]])
  title <- names(observation)[[id]]
  # View(quantile_ind)
  # View(observation_ind)
  
  plot <- ggplot() +
    geom_line(data = quantile_ind, aes(x = time, y = q05), 
              color = "dodgerblue", linetype = "dashed",
              alpha = 0.5) +
    geom_line(data = quantile_ind, aes(x = time, y = q95),
              color = "dodgerblue", linetype = "dashed",
              alpha = 0.5) +
    geom_ribbon(data = quantile_ind, 
                aes(x = time, ymin = q05, ymax = q95),
                fill = "dodgerblue", alpha = 0.5) +
    geom_line(data = quantile_ind, aes(x = time, y = q10), 
              color = "dodgerblue2", linetype = "dashed",
              alpha = 0.7) +
    geom_line(data = quantile_ind, aes(x = time, y = q90),
              color = "dodgerblue2", linetype = "dashed",
              alpha = 0.7) +
    geom_ribbon(data = quantile_ind, 
                aes(x = time, ymin = q10, ymax = q90),
                fill = "dodgerblue2", alpha = 0.7) +
    geom_line(data = quantile_ind, aes(x = time, y = q25), 
              color = "dodgerblue3", linetype = "dashed",
              alpha = 0.8) +
    geom_line(data = quantile_ind, aes(x = time, y = q75), 
              color = "dodgerblue3", linetype = "dashed",
              alpha = 0.8) +
    geom_ribbon(data = quantile_ind, 
                aes(x = time, ymin = q25, ymax = q75),
                fill = "dodgerblue3", alpha = 0.8) +
    geom_line(data = quantile_ind, aes(x = time, y = q50),
              color = "dodgerblue4", linetype = "dashed") +
    geom_point(data = observation_ind, aes(x = time, y = ht), 
               color = "black", size = 1) +
    theme_bw() +
    xlab("Time (yr)") +
    ylab("Height (cm)") +
    ggtitle(title) +
    xlim(0, max(observation_ind$time) + 1)
  
    # print(range(observation_ind$time))
    plot
}
```


```{r}
pdf("plot/people_like_me_all_plot_multiple_time_L2_norm_n15.pdf")
plot_all <- map(1:length(unique(train$id)), ~plm_plot(., quantile = result,
                              observation = dataset_all_long_ind))
plot_all
dev.off()
```


How to get teh mse....


squared distance to find the matches

means square error to show how good the model is
and matching trajectory plot 

add trajectory for brokenstick


```{r}
View(result)
View(train)
```
